+++ addOns/graphql/CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Fixed
+- A message is displayed if the "data" object in an introspection response
+  is `null` ([Issue 6890](https://github.com/zaproxy/zaproxy/issues/6890)).
 
 ## [0.6.0] - 2021-10-06
 ### Changed